ocr/requirements.txt
Django>=1.8,<2.0
pika
Pillow==3.3.0
gunicorn==19.6.0
opencv-python
pytesseract
